% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lab.note.R
\docType{package}
\name{lab.note}
\alias{lab.note}
\alias{lab.note-package}
\title{lab.note: Create Laboratory Note and Presentation Using R Markdown}
\description{
lab.note: Create Laboratory Note and Presentation Using R Markdown
}

